sheet_updater/requirements.txt
gspread
google-auth
google-generativeai
python-dotenv